booking service added

+++ src/main/java/com/irctc/booking/entity/Train.java
@@ -20,6 +20,6 @@ public class Train {
     private String  source;
     private String destination;
     private int totalSeats;
-    private double fare;
+    private float fare;
 
 }
